First pass on AlertManager v2
- Introduce legacy_alertmanager/grafana_alerting integrations
- Hanlde legacy/v2 integrations in public API
– Handle alerts from legacy/v2 integrations on same urls
