+++ engine/apps/alerts/integration_options_mixin.py
@@ -25,6 +25,8 @@ class IntegrationOptionsMixin:
     for integration_config in _config:
         vars()[f"INTEGRATION_{integration_config.slug.upper()}"] = integration_config.slug
 
+    INTEGRATION_TYPES = set(integration_config.slug for integration_config in _config)
+
     INTEGRATION_CHOICES = tuple(
         (
             (
@@ -39,7 +41,6 @@ class IntegrationOptionsMixin:
     WEB_INTEGRATION_CHOICES = [
         integration_config.slug for integration_config in _config if integration_config.is_displayed_on_web
     ]
-    PUBLIC_API_INTEGRATION_MAP = {integration_config.slug: integration_config.slug for integration_config in _config}
     INTEGRATIONS_TO_INSTRUCTIONS_WEB = {
         integration_config.slug: f"html/integration_{integration_config.slug}.html" for integration_config in _config
     }
