+++ engine/apps/integrations/metadata/heartbeat/alertmanager.py
@@ -1,9 +1,9 @@
 from pathlib import PurePath
 
-from apps.integrations.metadata.heartbeat._heartbeat_text_creator import HeartBeatTextCreatorForTitleGrouping
+from apps.integrations.metadata.heartbeat._heartbeat_text_creator import HeartBeatTextCreator
 
 integration_verbal = PurePath(__file__).stem
-creator = HeartBeatTextCreatorForTitleGrouping(integration_verbal)
+creator = HeartBeatTextCreator(integration_verbal)
 heartbeat_text = creator.get_heartbeat_texts()
 
 heartbeat_instruction_template = heartbeat_text.heartbeat_instruction_template
@@ -12,24 +12,65 @@ heartbeat_expired_title = heartbeat_text.heartbeat_expired_title
 heartbeat_expired_message = heartbeat_text.heartbeat_expired_message
 
 heartbeat_expired_payload = {
-    "endsAt": "",
-    "labels": {"alertname": heartbeat_expired_title},
+    "alerts": [
+        {
+            "endsAt": "",
+            "labels": {
+                "alertname": "OnCallHeartBeatMissing",
+            },
+            "status": "firing",
+            "startsAt": "",
+            "annotations": {
+                "title": heartbeat_expired_title,
+                "description": heartbeat_expired_message,
+            },
+            "fingerprint": "fingerprint",
+            "generatorURL": "",
+        },
+    ],
     "status": "firing",
-    "startsAt": "",
-    "annotations": {
-        "message": heartbeat_expired_message,
-    },
-    "generatorURL": None,
+    "version": "4",
+    "groupKey": '{}:{alertname="OnCallHeartBeatMissing"}',
+    "receiver": "",
+    "numFiring": 1,
+    "externalURL": "",
+    "groupLabels": {"alertname": "OnCallHeartBeatMissing"},
+    "numResolved": 0,
+    "commonLabels": {"alertname": "OnCallHeartBeatMissing"},
+    "truncatedAlerts": 0,
+    "commonAnnotations": {},
 }
 
 heartbeat_restored_title = heartbeat_text.heartbeat_restored_title
 heartbeat_restored_message = heartbeat_text.heartbeat_restored_message
 
+
 heartbeat_restored_payload = {
-    "endsAt": "",
-    "labels": {"alertname": heartbeat_restored_title},
-    "status": "resolved",
-    "startsAt": "",
-    "annotations": {"message": heartbeat_restored_message},
-    "generatorURL": None,
+    "alerts": [
+        {
+            "endsAt": "",
+            "labels": {
+                "alertname": "OnCallHeartBeatMissing",
+            },
+            "status": "resolved",
+            "startsAt": "",
+            "annotations": {
+                "title": heartbeat_restored_title,
+                "description": heartbeat_restored_message,
+            },
+            "fingerprint": "fingerprint",
+            "generatorURL": "",
+        },
+    ],
+    "status": "firing",
+    "version": "4",
+    "groupKey": '{}:{alertname="OnCallHeartBeatMissing"}',
+    "receiver": "",
+    "numFiring": 0,
+    "externalURL": "",
+    "groupLabels": {"alertname": "OnCallHeartBeatMissing"},
+    "numResolved": 1,
+    "commonLabels": {"alertname": "OnCallHeartBeatMissing"},
+    "truncatedAlerts": 0,
+    "commonAnnotations": {},
 }
