+++ engine/apps/integrations/templates/heartbeat_instructions/legacy_alertmanager.html
@@ -1,0 +1,41 @@
+<p>This configuration will send an alert once a minute, and if alertmanager stops working, OnCall will detect
+    it and notify you about that.</p>
+<ol>
+    <li>
+        <p>Add the alert generating script to <code>prometheus.yaml</code> file.
+        Within Prometheus it is trivial to create an expression that we can use as a heartbeat for OnCall,
+        like <code>vector(1)</code>. That expression will always return true.</p>
+        <p>Here is an alert that leverages the previous expression to create a heartbeat alert:</p>
+        <pre><code>
+            groups:
+            - name: meta
+              rules:
+              - alert: heartbeat
+                expr: vector(1)
+                labels:
+                  severity: none
+                annotations:
+                  description: This is a heartbeat alert for Grafana OnCall
+                  summary: Heartbeat for Grafana OnCall
+        </code></pre>
+    </li>
+    <li><p>Add receiver configuration to <code>prometheus.yaml</code> with the unique url from OnCall global:</p>
+        <pre><code>
+            ...
+            route:
+            ...
+                routes:
+                - match:
+                    alertname: heartbeat
+                  receiver: 'grafana-oncall-heartbeat'
+                  group_wait: 0s
+                  group_interval: 1m
+                  repeat_interval: 50s
+            receivers:
+            - name: 'grafana-oncall-heartbeat'
+            webhook_configs:
+            - url: {{ heartbeat_url }}
+                send_resolved: false
+        </code></pre>
+    </li>
+</ol>
