+++ engine/apps/integrations/templates/html/integration_legacy_grafana_alerting.html
@@ -1,0 +1,62 @@
+<h4>Congratulations, you've connected the Grafana Alerting and Grafana OnCall!</h4>
+<blockquote>
+    This is the integration with current Grafana Alerting.
+    It already automatically created a new Grafana Alerting <code class='code-inline'>Contact Point</code> and
+    a <code class='code-inline'>Specific Route</code>.<br>
+    If you want to connect the other Grafana Instance please
+    choose the <code class='code-inline'>Other Grafana</code> Integration instead.
+</blockquote>
+
+<h4>How to send the Test alert from Grafana Alerting?</h4>
+<p>
+<ol>
+    <li>
+        1. Open the corresponding Grafana Alerting <code class='code-inline'>Contact Point</code>
+    </li>
+    <li>
+        2. Use the <code class='code-inline'>Test</code> buton to send an alert to Grafana OnCall
+    </li>
+</ol>
+</p>
+
+<h4>How to choose what alerts to send from Grafana Alerting to Grafana OnCall?</h4>
+<p>
+<ol>
+    <li>
+        1. Open the corresponding Grafana Alerting <code class='code-inline'>Specific Route</code>
+    </li>
+    <li>
+        2. All alerts are sent from Grafana Alerting to Grafana OnCall by default,
+        specify Matching Labels to select which alerts to send
+    </li>
+</ol>
+</p>
+
+<h4>What if the Grafana Alerting <code class='code-inline'>Contact Point</code> is missing?</h4>
+<p>
+<ol>
+    <li>
+        1. May be it was deleted, you can always re-create them manually
+    </li>
+    <li>
+        2. Use the following webhook url to create a webhook
+        <code class='code-inline'>Contact Point</code> in Grafana Alerting
+        <pre>{{ alert_receive_channel.integration_url }}</pre>
+    </li>
+</ol>
+</p>
+
+<h4>Next steps:</h4>
+<p><ol>
+    <li>
+        1. Add the routes and escalations in <code class='code-inline'>Escalations settings</code>
+    </li>
+    <li>
+        2. Check grouping, auto-resolving, and rendering templates in
+        <code class='code-inline'>Alert Templates</code> Settings
+    </li>
+    <li>
+        3. Make sure all the users set up their <code class='code-inline'>Personal Notifications</code> Settings
+        on the <code class='code-inline'>Users</code> Page
+    </li>
+</ol></p>
