+++ engine/apps/integrations/urls.py
@@ -8,7 +8,6 @@ from common.api_helpers.optional_slash_router import optional_slash_path
 
 from .views import (
     AlertManagerAPIView,
-    AlertManagerV2View,
     AmazonSNS,
     GrafanaAlertingAPIView,
     GrafanaAPIView,
@@ -32,7 +31,6 @@ urlpatterns = [
     path("grafana_alerting/<str:alert_channel_key>/", GrafanaAlertingAPIView.as_view(), name="grafana_alerting"),
     path("alertmanager/<str:alert_channel_key>/", AlertManagerAPIView.as_view(), name="alertmanager"),
     path("amazon_sns/<str:alert_channel_key>/", AmazonSNS.as_view(), name="amazon_sns"),
-    path("alertmanager_v2/<str:alert_channel_key>/", AlertManagerV2View.as_view(), name="alertmanager_v2"),
     path("<str:integration_type>/<str:alert_channel_key>/", UniversalAPIView.as_view(), name="universal"),
 ]
