+++ engine/apps/integrations/views.py
@@ -99,11 +99,23 @@ class AlertManagerAPIView(
         """
         alert_receive_channel = self.request.alert_receive_channel
         if not self.check_integration_type(alert_receive_channel):
+            print("BOLO")
             return HttpResponseBadRequest(
                 f"This url is for integration with {alert_receive_channel.get_integration_display()}. Key is for "
                 + str(alert_receive_channel.get_integration_display())
             )
 
+        if alert_receive_channel.is_legacy:
+            self.process_v1(request, alert_receive_channel)
+        else:
+            self.process_v2(request, alert_receive_channel)
+
+        return Response("Ok.")
+
+    def process_v1(self, request, alert_receive_channel):
+        """
+        process_v1 creates alerts from each alert in incoming AlertManager payload.
+        """
         for alert in request.data.get("alerts", []):
             if settings.DEBUG:
                 create_alertmanager_alerts(alert_receive_channel.pk, alert)
@@ -115,17 +127,47 @@ class AlertManagerAPIView(
 
                 create_alertmanager_alerts.apply_async((alert_receive_channel.pk, alert))
 
-        return Response("Ok.")
+    def process_v2(self, request, alert_receive_channel):
+        """
+        process_v2 creates one alert from one incoming AlertManager payload
+        """
+        alerts = request.data.get("alerts", [])
+
+        data = request.data
+        if "firingAlerts" not in request.data:
+            # Count firing and resolved alerts manually if not present in payload
+            num_firing = len(list(filter(lambda a: a["status"] == "firing", alerts)))
+            num_resolved = len(list(filter(lambda a: a["status"] == "resolved", alerts)))
+            data = {**request.data, "firingAlerts": num_firing, "resolvedAlerts": num_resolved}
+
+        create_alert.apply_async(
+            [],
+            {
+                "title": None,
+                "message": None,
+                "image_url": None,
+                "link_to_upstream_details": None,
+                "alert_receive_channel_pk": alert_receive_channel.pk,
+                "integration_unique_data": None,
+                "raw_request_data": data,
+            },
+        )
 
     def check_integration_type(self, alert_receive_channel):
-        return alert_receive_channel.integration == AlertReceiveChannel.INTEGRATION_ALERTMANAGER
+        return alert_receive_channel.integration in {
+            AlertReceiveChannel.INTEGRATION_ALERTMANAGER,
+            AlertReceiveChannel.INTEGRATION_LEGACY_ALERTMANAGER,
+        }
 
 
 class GrafanaAlertingAPIView(AlertManagerAPIView):
     """Grafana Alerting has the same payload structure as AlertManager"""
 
     def check_integration_type(self, alert_receive_channel):
-        return alert_receive_channel.integration == AlertReceiveChannel.INTEGRATION_GRAFANA_ALERTING
+        return alert_receive_channel.integration in {
+            AlertReceiveChannel.INTEGRATION_GRAFANA_ALERTING,
+            AlertReceiveChannel.INTEGRATION_LEGACGRAFANA_ALERTING,
+        }
 
 
 class GrafanaAPIView(AlertManagerAPIView):
@@ -270,46 +312,3 @@ class IntegrationHeartBeatAPIView(AlertChannelDefiningMixin, IntegrationHeartBea
         process_heartbeat_task.apply_async(
             (alert_receive_channel.pk,),
         )
-
-
-class AlertManagerV2View(BrowsableInstructionMixin, AlertChannelDefiningMixin, IntegrationRateLimitMixin, APIView):
-    """
-    AlertManagerV2View consumes alerts from AlertManager. It expects data to be in format of AM webhook receiver.
-    """
-
-    def post(self, request, *args, **kwargs):
-        alert_receive_channel = self.request.alert_receive_channel
-        if not alert_receive_channel.integration == AlertReceiveChannel.INTEGRATION_ALERTMANAGER_V2:
-            return HttpResponseBadRequest(
-                f"This url is for integration with {alert_receive_channel.config.title}."
-                f"Key is for {alert_receive_channel.get_integration_display()}"
-            )
-        alerts = request.data.get("alerts", [])
-
-        data = request.data
-        if "numFiring" not in request.data:
-            num_firing = 0
-            num_resolved = 0
-            for a in alerts:
-                if a["status"] == "firing":
-                    num_firing += 1
-                elif a["status"] == "resolved":
-                    num_resolved += 1
-            # Count firing and resolved alerts manually if not present in payload
-            data = {**request.data, "numFiring": num_firing, "numResolved": num_resolved}
-        else:
-            data = request.data
-
-        create_alert.apply_async(
-            [],
-            {
-                "title": None,
-                "message": None,
-                "image_url": None,
-                "link_to_upstream_details": None,
-                "alert_receive_channel_pk": alert_receive_channel.pk,
-                "integration_unique_data": None,
-                "raw_request_data": data,
-            },
-        )
-        return Response("Ok.")
