+++ engine/config_integrations/grafana.py
@@ -8,8 +8,8 @@ is_displayed_on_web = True
 is_featured = False
 is_able_to_autoresolve = True
 is_demo_alert_enabled = True
+based_on_am = True
 
-description = None
 
 # Default templates
 slack_title = """\
