+++ engine/config_integrations/grafana_alerting.py
@@ -12,120 +12,272 @@ featured_tag_name = "Quick Connect"
 is_able_to_autoresolve = True
 is_demo_alert_enabled = True
 
-description = """ \
-Alerts from Grafana Alertmanager are automatically routed to this integration.
-{% for dict_item in grafana_alerting_entities %}
-<br>Click <a href='{{dict_item.contact_point_url}}' target='_blank'>here</a>
- to open contact point, and
- <a href='{{dict_item.routes_url}}' target='_blank'>here</a>
- to open Notification policy for {{dict_item.alertmanager_name}} Alertmanager.
-{% endfor %}
-{% if not is_finished_alerting_setup %}
-<br>Creating contact points and routes for other alertmanagers...
+
+# Behaviour
+source_link = "{{ payload.externalURL }}"
+
+grouping_id = "{{ payload.groupKey }}"
+
+resolve_condition = """{{ payload.status == "resolved" }}"""
+
+acknowledge_condition = None
+
+
+web_title = """\
+{%- set groupLabels = payload.groupLabels.copy() -%}
+{%- set alertname = groupLabels.pop('alertname') | default("") -%}
+
+
+[{{ payload.status }}{% if payload.status == 'firing' %}:{{ payload.numFiring }}{% endif %}] {{ alertname }} {% if groupLabels | length > 0 %}({{ groupLabels|join(", ") }}){% endif %}
+"""  # noqa
+
+web_message = """\
+{%- set annotations = payload.commonAnnotations.copy() -%}
+
+{% set severity = payload.groupLabels.severity -%}
+{% if severity %}
+{%- set severity_emoji = {"critical": ":rotating_light:", "warning": ":warning:" }[severity] | default(":question:") -%}
+Severity: {{ severity }} {{ severity_emoji }}
 {% endif %}
+
+{%- set status = payload.status | default("Unknown") %}
+{%- set status_emoji = {"firing": ":fire:", "resolved": ":white_check_mark:"}[status] | default(":warning:") %}
+Status: {{ status }} {{ status_emoji }} (on the source)
+{% if status == "firing" %}
+Firing alerts – {{ payload.numFiring }}
+Resolved alerts – {{ payload.numResolved }}
+{% endif %}
+
+{% if "runbook_url" in annotations -%}
+[:book: Runbook:link:]({{ annotations.runbook_url }})
+{%- set _ = annotations.pop('runbook_url') -%}
+{%- endif %}
+
+{%- if "runbook_url_internal" in annotations -%}
+[:closed_book: Runbook (internal):link:]({{ annotations.runbook_url_internal }})
+{%- set _ = annotations.pop('runbook_url_internal') -%}
+{%- endif %}
+
+GroupLabels:
+{%- for k, v in payload["groupLabels"].items() %}
+- {{ k }}: {{ v }}
+{%- endfor %}
+
+{% if payload["commonLabels"] | length > 0 -%}
+CommonLabels:
+{%- for k, v in payload["commonLabels"].items() %}
+- {{ k }}: {{ v }}
+{%- endfor %}
+{% endif %}
+
+{% if annotations | length > 0 -%}
+Annotations:
+{%- for k, v in annotations.items() %}
+- {{ k }}: {{ v }}
+{%- endfor %}
+{% endif %}
+
+[View in AlertManager]({{ source_link }})
 """
 
-# Default templates
+
+# Slack templates
 slack_title = """\
-{# Usually title is located in payload.labels.alertname #}
-{% set title = payload.get("labels", {}).get("alertname", "No title (check Web Title Template)") %}
-{# Combine the title from different built-in variables into slack-formatted url #}
-*<{{ grafana_oncall_link }}|#{{ grafana_oncall_incident_id }} {{ title }}>* via {{ integration_name }}
+{%- set groupLabels = payload.groupLabels.copy() -%}
+{%- set alertname = groupLabels.pop('alertname') | default("") -%}
+*<{{ grafana_oncall_link }}|#{{ grafana_oncall_incident_id }} {{ web_title }}>* via {{ integration_name }}
 {% if source_link %}
  (*<{{ source_link }}|source>*)
 {%- endif %}
 """
 
+# default slack message template is identical to web message template, except urls
+# It can be based on web message template (see example), but it can affect existing templates
+# slack_message = """
+# {% set mkdwn_link_regex = "\[([\w\s\d:]+)\]\((https?:\/\/[\w\d./?=#]+)\)" %}
+# {{ web_message
+#   | regex_replace(mkdwn_link_regex, "<\\2|\\1>")
+# }}
+# """
+
 slack_message = """\
-{{- payload.message }}
-{%- if "status" in payload -%}
-*Status*: {{ payload.status }}
-{% endif -%}
-*Labels:* {% for k, v in payload["labels"].items() %}
-{{ k }}: {{ v }}{% endfor %}
-*Annotations:* 
-{%- for k, v in payload.get("annotations", {}).items() %}
-{#- render annotation as slack markdown url if it starts with http #}
-{{ k }}: {% if v.startswith("http") %} <{{v}}|here> {% else %} {{v}} {% endif -%}
-{% endfor %}
-"""  # noqa:W291
+{%- set annotations = payload.commonAnnotations.copy() -%}
+
+{% set severity = payload.groupLabels.severity -%}
+{% if severity %}
+{%- set severity_emoji = {"critical": ":rotating_light:", "warning": ":warning:" }[severity] | default(":question:") -%}
+Severity: {{ severity }} {{ severity_emoji }}
+{% endif %}
+
+{%- set status = payload.status | default("Unknown") %}
+{%- set status_emoji = {"firing": ":fire:", "resolved": ":white_check_mark:"}[status] | default(":warning:") %}
+Status: {{ status }} {{ status_emoji }} (on the source)
+{% if status == "firing" %}
+Firing alerts – {{ payload.numFiring }}
+Resolved alerts – {{ payload.numResolved }}
+{% endif %}
+
+{% if "runbook_url" in annotations -%}
+<{{ annotations.runbook_url }}|:book: Runbook:link:>
+{%- set _ = annotations.pop('runbook_url') -%}
+{%- endif %}
+
+{%- if "runbook_url_internal" in annotations -%}
+<{{ annotations.runbook_url_internal }}|:closed_book: Runbook (internal):link:>
+{%- set _ = annotations.pop('runbook_url_internal') -%}
+{%- endif %}
+
+GroupLabels:
+{%- for k, v in payload["groupLabels"].items() %}
+- {{ k }}: {{ v }}
+{%- endfor %}
+
+{% if payload["commonLabels"] | length > 0 -%}
+CommonLabels:
+{%- for k, v in payload["commonLabels"].items() %}
+- {{ k }}: {{ v }}
+{%- endfor %}
+{% endif %}
+
+{% if annotations | length > 0 -%}
+Annotations:
+{%- for k, v in annotations.items() %}
+- {{ k }}: {{ v }}
+{%- endfor %}
+{% endif %}
+"""
+# noqa: W291
 
 
 slack_image_url = None
 
-web_title = """\
-{# Usually title is located in payload.labels.alertname #}
-{{- payload.get("labels", {}).get("alertname", "No title (check Web Title Template)") }}
-"""
+web_image_url = None
 
-web_message = """\
-{{- payload.message }}
-{%- if "status" in payload %}
-**Status**: {{ payload.status }}
-{% endif -%}
-**Labels:** {% for k, v in payload["labels"].items() %}
-*{{ k }}*: {{ v }}{% endfor %}
-**Annotations:** 
-{%- for k, v in payload.get("annotations", {}).items() %}
-{#- render annotation as markdown url if it starts with http #}
-*{{ k }}*: {% if v.startswith("http") %} [here]({{v}}){% else %} {{v}} {% endif -%}
-{% endfor %}
-"""  # noqa:W291
+sms_title = web_title
 
 
-web_image_url = slack_image_url
+phone_call_title = """{{ payload.groupLabels|join(", ") }}"""
 
-sms_title = '{{ payload.get("labels", {}).get("alertname", "Title undefined") }}'
-phone_call_title = sms_title
-
-telegram_title = sms_title
+telegram_title = web_title
 
 telegram_message = """\
-{{- payload.messsage }}
-{%- if "status" in payload -%}
-<b>Status</b>: {{ payload.status }}
-{% endif -%}
-<b>Labels:</b> {% for k, v in payload["labels"].items() %}
-{{ k }}: {{ v }}{% endfor %}
-<b>Annotations:</b> 
-{%- for k, v in payload.get("annotations", {}).items() %}
-{#- render annotation as markdown url if it starts with http #}
-{{ k }}: {{ v }}
-{% endfor %}"""  # noqa:W291
+{%- set annotations = payload.commonAnnotations.copy() -%}
 
-telegram_image_url = slack_image_url
+{% set severity = payload.groupLabels.severity -%}
+{% if severity %}
+{%- set severity_emoji = {"critical": ":rotating_light:", "warning": ":warning:" }[severity] | default(":question:") -%}
+Severity: {{ severity }} {{ severity_emoji }}
+{% endif %}
 
-source_link = "{{ payload.generatorURL }}"
+{%- set status = payload.status | default("Unknown") %}
+{%- set status_emoji = {"firing": ":fire:", "resolved": ":white_check_mark:"}[status] | default(":warning:") %}
+Status: {{ status }} {{ status_emoji }} (on the source)
+{% if status == "firing" %}
+Firing alerts – {{ payload.numFiring }}
+Resolved alerts – {{ payload.numResolved }}
+{% endif %}
 
-grouping_id = web_title
+{% if "runbook_url" in annotations -%}
+<a href='{{ annotations.runbook_url }}'>:book: Runbook:link:</a>
+{%- set _ = annotations.pop('runbook_url') -%}
+{%- endif %}
 
-resolve_condition = """\
-{{ payload.get("status", "") == "resolved" }}
+{%- if "runbook_url_internal" in annotations -%}
+<a href='{{ annotations.runbook_url_internal }}'>:closed_book: Runbook (internal):link:</a>
+{%- set _ = annotations.pop('runbook_url_internal') -%}
+{%- endif %}
+
+GroupLabels:
+{%- for k, v in payload["groupLabels"].items() %}
+- {{ k }}: {{ v }}
+{%- endfor %}
+
+{% if payload["commonLabels"] | length > 0 -%}
+CommonLabels:
+{%- for k, v in payload["commonLabels"].items() %}
+- {{ k }}: {{ v }}
+{%- endfor %}
+{% endif %}
+
+{% if annotations | length > 0 -%}
+Annotations:
+{%- for k, v in annotations.items() %}
+- {{ k }}: {{ v }}
+{%- endfor %}
+{% endif %}
+
+<a href='{{ source_link }}'>View in AlertManager</a>
 """
 
-acknowledge_condition = None
+telegram_image_url = None
+
 
 example_payload = {
-    "receiver": "amixr",
-    "status": "firing",
     "alerts": [
         {
-            "status": "firing",
-            "labels": {
-                "alertname": "TestAlert",
-                "region": "eu-1",
-            },
-            "annotations": {"description": "This alert was sent by user for demonstration purposes"},
-            "startsAt": "2018-12-25T15:47:47.377363608Z",
             "endsAt": "0001-01-01T00:00:00Z",
+            "labels": {
+                "job": "node",
+                "group": "production",
+                "instance": "localhost:8081",
+                "severity": "critical",
+                "alertname": "InstanceDown",
+            },
+            "status": "firing",
+            "startsAt": "2023-06-12T08:24:38.326Z",
+            "annotations": {
+                "title": "Instance localhost:8081 down",
+                "description": "localhost:8081 of job node has been down for more than 1 minute.",
+            },
+            "fingerprint": "f404ecabc8dd5cd7",
             "generatorURL": "",
-            "amixr_demo": True,
-        }
+        },
+        {
+            "endsAt": "0001-01-01T00:00:00Z",
+            "labels": {
+                "job": "node",
+                "group": "canary",
+                "instance": "localhost:8082",
+                "severity": "critical",
+                "alertname": "InstanceDown",
+            },
+            "status": "firing",
+            "startsAt": "2023-06-12T08:24:38.326Z",
+            "annotations": {
+                "title": "Instance localhost:8082 down",
+                "description": "localhost:8082 of job node has been down for more than 1 minute.",
+            },
+            "fingerprint": "f8f08d4e32c61a9d",
+            "generatorURL": "",
+        },
+        {
+            "endsAt": "0001-01-01T00:00:00Z",
+            "labels": {
+                "job": "node",
+                "group": "production",
+                "instance": "localhost:8083",
+                "severity": "critical",
+                "alertname": "InstanceDown",
+            },
+            "status": "firing",
+            "startsAt": "2023-06-12T08:24:38.326Z",
+            "annotations": {
+                "title": "Instance localhost:8083 down",
+                "description": "localhost:8083 of job node has been down for more than 1 minute.",
+            },
+            "fingerprint": "39f38c0611ee7abd",
+            "generatorURL": "",
+        },
     ],
-    "groupLabels": {},
-    "commonLabels": {},
-    "commonAnnotations": {},
-    "externalURL": "http://f1d1ef51d710:9093",
+    "status": "firing",
     "version": "4",
-    "groupKey": "{}:{}",
+    "groupKey": '{}:{alertname="InstanceDown"}',
+    "receiver": "combo",
+    "numFiring": 3,
+    "externalURL": "",
+    "groupLabels": {"alertname": "InstanceDown"},
+    "numResolved": 0,
+    "commonLabels": {"job": "node", "severity": "critical", "alertname": "InstanceDown"},
+    "truncatedAlerts": 0,
+    "commonAnnotations": {},
 }
